抽离独立的自动补全脚本
