+++ src/commands/completion-install.ts
@@ -1,4 +1,5 @@
 import Command from '@oclif/command';
+import { resolve } from 'path';
 import * as tabtab from 'tabtab';
 
 export default class InstallCompletion extends Command {
@@ -7,9 +8,11 @@ export default class InstallCompletion extends Command {
   static args = [];
 
   async run() {
+    const node = process.argv[0];
+    const script = resolve(__dirname, '../../lib/completion.js');
     await tabtab.install({
       name: 'lisa',
-      completer: 'lisa',
+      completer: `${node} ${script}`,
     });
   }
 }
